Fix real devkitA64 build failure: no dlfcn.h on Switch (native_glad.c)
The first real devkitA64/libnx CI build (previous commit was only checked
against a local x86_64 syntax-check harness with hand-written stub
headers) failed with:

  platform/native_glad.c:95:10: fatal error: dlfcn.h: No such file or
  directory

native_glad.c's non-Windows glad loader branch unconditionally
`#include <dlfcn.h>` and calls dlopen/dlsym/dlclose to load libGL and
resolve GL entry points at runtime. That header/API does not exist on
Vita (newlib) or on Switch (devkitA64/libnx) — neither platform loads a
shared libGL; both resolve GL entry points directly through a
platform-specific getter (vglGetProcAddress on Vita,
NativeRendererSwitch_GetProcAddress on Switch, already used a few lines
down in get_proc()). The file already carved out __vita__ as a no-op in
open_gl()/close_gl(), but never got the same treatment for __SWITCH__,
so the #include <dlfcn.h> and the raw dlopen/dlsym/dlclose calls it
guards were still compiled in for Switch.

Fix: extend the existing __vita__ no-op guards in native_glad.c to
__SWITCH__ as well (both for the #include <dlfcn.h> and for the
dlopen/gladGetProcAddressPtr/dlclose code paths). No behavior change for
Vita/PC.

Also fixed a related latent -Wformat bug the CI log surfaced in
native_gpu.c: a debug-log NATIVE_GPU_ERROR() call formatted a pointer
difference (ptrdiff_t, 8 bytes on LP64/AArch64) with %d; changed to cast
to long and use %ld so the varargs match on both ILP32 and LP64 targets.

Verified with the local x86_64 -m64 syntax-check harness:
-Wpointer-to-int-cast -Wint-to-pointer-cast -Wincompatible-pointer-types
-> 0 warnings, 0 errors (same as before this fix; this pass targets the
dlfcn.h/build-error class of problem the real toolchain caught that the
syntax-only harness cannot, since it links against a real dlfcn.h that
Switch's toolchain does not ship).

+++ platform/native_gpu.c
@@ -2787,8 +2787,12 @@ void ParsePrimitivesLinkedList(u32 *p, int singlePrimitive)
 				{
 					char packetRegion[64];
 					NativeGpu_FormatPointerRegion(packetRegion, sizeof(packetRegion), (uintptr_t)basePacket);
-					NATIVE_GPU_ERROR("did not output valid primitive or ptag length is not valid (diff=%d packet=%p region=%s words=%08x %08x %08x %08x)\n",
-					                 endPacket - currentPacket, (void *)basePacket, packetRegion, NativeGpu_ReadPacketWordForLog((uintptr_t)basePacket, 0),
+					// NOTE: pointer difference is ptrdiff_t (widens to 8 bytes
+					// on LP64 targets like Switch/AArch64); cast to long and
+					// format with %ld instead of %d to avoid a varargs
+					// type/format mismatch across platforms.
+					NATIVE_GPU_ERROR("did not output valid primitive or ptag length is not valid (diff=%ld packet=%p region=%s words=%08x %08x %08x %08x)\n",
+					                 (long)(endPacket - currentPacket), (void *)basePacket, packetRegion, NativeGpu_ReadPacketWordForLog((uintptr_t)basePacket, 0),
 					                 NativeGpu_ReadPacketWordForLog((uintptr_t)basePacket, 1), NativeGpu_ReadPacketWordForLog((uintptr_t)basePacket, 2),
 					                 NativeGpu_ReadPacketWordForLog((uintptr_t)basePacket, 3));
 				}
